Move test files next to tested files
