+++ p2panda-js/src/instance/instance.test.ts
@@ -3,9 +3,9 @@
 import { expect } from 'chai';
 
 // @ts-expect-error bundle import has no type
-import { Session } from '../lib';
-import { getMessageFields } from '../src/instance';
-import { marshallRequestFields } from '../src/utils';
+import { Session } from '~/../lib';
+import { getMessageFields } from '.';
+import { marshallRequestFields } from '~/utils';
 
 describe('Instance tools', () => {
   describe('getMessageFields', () => {
