+++ p2panda-js/src/wasm.test.ts
@@ -3,7 +3,7 @@
 import { expect } from 'chai';
 
 // @ts-expect-error bundle import has no type
-import { wasm } from '../lib';
+import { wasm } from '~/../lib';
 
 describe('Web assembly interface', () => {
   describe('KeyPair', () => {
